just separate file && norm
