+++ srcs/builtin/ft_cd_canonicalize.c
@@ -1,53 +1,9 @@
-#include <stdlib.h>
 #include "minishell.h"
 #include "ms_builtin.h"
 #include "ft_deque.h"
-#include "ft_dprintf.h"
 #include "ft_mem.h"
 #include "ft_string.h"
 #include "ft_sys.h"
-
-static void	add_split_path_to_path_elems(t_deque *path_elems, char **split_path)
-{
-	size_t			i;
-	t_deque_node	*node;
-
-	i = 0;
-	while (split_path[i])
-	{
-		node = deque_node_new((void *)split_path[i]);
-		deque_add_back(path_elems, node);
-		split_path[i] = NULL;
-		i++;
-	}
-}
-
-static void	split_and_add_dq(t_deque *path_elems, const char *path)
-{
-	char	**split_path;
-
-	split_path = ft_split(path, PATH_DELIMITER_CHR);
-	if (!split_path)
-		ft_abort();
-	add_split_path_to_path_elems(path_elems, split_path);
-	ft_free(&split_path);
-}
-
-static t_deque	*separate_path_and_join(const char *path, t_context *context)
-{
-	t_deque	*path_elems;
-
-	path_elems = deque_new();
-	if (!path_elems)
-		ft_abort();
-	if (!is_absolute_path(path))
-	{
-		if (context->internal_pwd)
-			split_and_add_dq(path_elems, context->internal_pwd);
-	}
-	split_and_add_dq(path_elems, path);
-	return (path_elems);
-}
 
 static size_t	calc_len_path(t_deque *path_elems)
 {
@@ -98,82 +54,9 @@ static char	*convert_path_elems_to_absolute_path(t_deque *path_elems)
 	return (absolute_path);
 }
 
-// todo
-static void	del(void *content)
+void	del_path_elem(void *content)
 {
 	ft_free(&content);
-}
-
-static void	erase_dot_path(t_deque **dq)
-{
-	t_deque			*stack;
-	t_deque_node	*pop_node;
-	char			*path_elem;
-
-	stack = deque_new();
-	if (!stack)
-		ft_abort();
-	while (!deque_is_empty(*dq))
-	{
-		pop_node = deque_pop_front(*dq);
-		path_elem = (char *)pop_node->content;
-		if (ft_streq(path_elem, PATH_DOT))
-		{
-			deque_clear_node(&pop_node, del);
-			continue ;
-		}
-		deque_add_back(stack, pop_node);
-	}
-	deque_clear_all(dq, del);
-	*dq = stack;
-}
-
-static void	erase_dot_dot_path(t_deque **dq)
-{
-	t_deque			*stack;
-	t_deque_node	*pop_node;
-	char			*path_elem;
-
-	stack = deque_new();
-	if (!stack)
-		ft_abort();
-	while (!deque_is_empty(*dq))
-	{
-		pop_node = deque_pop_front(*dq);
-		path_elem = (char *)pop_node->content;
-		if (ft_streq(path_elem, PATH_DOT_DOT))
-		{
-			deque_clear_node(&pop_node, del);
-			pop_node = deque_pop_back(stack);
-			deque_clear_node(&pop_node, del);
-			continue ;
-		}
-		deque_add_back(stack, pop_node);
-	}
-	deque_clear_all(dq, del);
-	*dq = stack;
-}
-
-static bool	is_head_double_slash(const char *path)
-{
-	return (path[0] == PATH_DELIMITER_CHR \
-			&& path[1] == PATH_DELIMITER_CHR \
-			&& path[2] != PATH_DELIMITER_CHR);
-}
-
-static char	*handle_double_slash_path(const char *path, char *absolute_path)
-{
-	char	*new_path;
-
-	if (is_head_double_slash(path))
-	{
-		new_path = ft_strjoin(PATH_DELIMITER_STR, absolute_path);
-		if (!new_path)
-			ft_abort();
-		ft_free(&absolute_path);
-		return (new_path);
-	}
-	return (absolute_path);
 }
 
 // 	 PWD         path
@@ -196,6 +79,6 @@ char	*cd_canonicalize_path(const char *path, t_context *context)
 	erase_dot_dot_path(&path_elems);
 	absolute_path = convert_path_elems_to_absolute_path(path_elems);
 	absolute_path = handle_double_slash_path(path, absolute_path);
-	deque_clear_all(&path_elems, del);
+	deque_clear_all(&path_elems, del_path_elem);
 	return (absolute_path);
 }
