+++ srcs/builtin/ft_cd_chdir.c
@@ -1,4 +1,3 @@
-#include <stdlib.h>
 #include <unistd.h>
 #include "minishell.h"
 #include "ms_builtin.h"
